+++ examples/quickstart/main.go
@@ -48,7 +48,10 @@ func (m *GreeterModule) Init(ctx context.Context, reg modulex.Registry) error {
 
 func main() {
 	logger := slog.New(slog.NewTextHandler(io.Discard, nil))
-	manager := modulex.NewManager(nil, logger, nil)
+	manager, err := modulex.NewManager(nil, logger, nil)
+	if err != nil {
+		panic(err)
+	}
 
 	if err := manager.RegisterModule(&LoggerModule{}); err != nil {
 		panic(err)